[refactor] 파일 이름 변경
구조를 간소화 하기위해 auth -> user로 바꾸었고, APIError에서 APIErrorInfos로 이름 변경

#74
